# docs/architecture/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveIn": {
        "properties": {
          "accept": {
            "title": "Accept",
            "type": "boolean"
          },
          "relationship": {
            "default": "guardian",
            "enum": [
              "mother",
              "father",
              "guardian",
              "other"
            ],
            "title": "Relationship",
            "type": "string"
          }
        },
        "required": [
          "accept"
        ],
        "title": "ApproveIn",
        "type": "object"
      },
...
      "ChildOut": {
        "properties": {
          "class_number": {
            "title": "Class Number",
            "type": "integer"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "full_name",
          "class_number",
          "status"
        ],
        "title": "ChildOut",
        "type": "object"
      },
...
      "ConsentLinkOut": {
        "properties": {
          "child_first_name": {
            "title": "Child First Name",
            "type": "string"
          },
          "class_label": {
            "title": "Class Label",
            "type": "string"
          },
          "consent_text": {
            "title": "Consent Text",
            "type": "string"
          },
          "consent_version": {
            "title": "Consent Version",
            "type": "string"
          }
        },
        "required": [
          "child_first_name",
          "class_label",
          "consent_version",
          "consent_text"
        ],
        "title": "ConsentLinkOut",
        "type": "object"
      },
...
      "LoginIn": {
        "properties": {
          "identifier": {
            "maxLength": 254,
            "minLength": 1,
            "title": "Identifier",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "identifier",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
...
      "MeOut": {
        "properties": {
          "account_type": {
            "title": "Account Type",
            "type": "string"
          },
          "children": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ChildOut"
            },
            "title": "Children",
            "type": "array"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 254,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "email_verified": {
            "title": "Email Verified",
            "type": "boolean"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mobile": {
            "anyOf": [
              {
                "maxLength": 254,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile"
          },
          "mobile_verified": {
            "title": "Mobile Verified",
            "type": "boolean"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "roles": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RoleOut"
            },
            "title": "Roles",
            "type": "array"
          },
          "student": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudentOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "full_name",
          "email_verified",
          "mobile_verified",
          "account_type"
        ],
        "title": "MeOut",
        "type": "object"
      },
...
      "OkOut": {
        "properties": {
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkOut",
        "type": "object"
      },
      "OtpSendIn": {
        "properties": {
          "destination": {
            "maxLength": 254,
            "minLength": 1,
            "title": "Destination",
            "type": "string"
          },
          "purpose": {
            "enum": [
              "verify_contact",
              "login",
              "reset_password"
            ],
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "destination",
          "purpose"
        ],
        "title": "OtpSendIn",
        "type": "object"
      },
      "OtpVerifyIn": {
        "properties": {
          "code": {
            "maxLength": 10,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "destination": {
            "maxLength": 254,
            "minLength": 1,
            "title": "Destination",
            "type": "string"
          },
          "purpose": {
            "enum": [
              "verify_contact",
              "login",
              "reset_password"
            ],
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "destination",
          "purpose",
          "code"
        ],
        "title": "OtpVerifyIn",
        "type": "object"
      },
      "ParentContactIn": {
        "properties": {
          "parent_contact": {
            "maxLength": 254,
            "minLength": 1,
            "title": "Parent Contact",
            "type": "string"
          }
        },
        "required": [
          "parent_contact"
        ],
        "title": "ParentContactIn",
        "type": "object"
      },
...
      "ParentSignupIn": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "maxLength": 254,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "maxLength": 150,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "mobile": {
            "anyOf": [
              {
                "maxLength": 254,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "preferred_language": {
            "default": "en",
            "maxLength": 10,
            "minLength": 2,
            "title": "Preferred Language",
            "type": "string"
          }
        },
        "required": [
          "full_name",
          "password"
        ],
        "title": "ParentSignupIn",
        "type": "object"
      },
      "PasswordResetIn": {
        "properties": {
          "code": {
            "maxLength": 10,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "destination": {
            "maxLength": 254,
            "minLength": 1,
            "title": "Destination",
            "type": "string"
          },
          "new_password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "destination",
          "code",
          "new_password"
        ],
        "title": "PasswordResetIn",
        "type": "object"
      },
...
      "RecordOut": {
        "properties": {
          "completed_courses": {
            "items": {
              "$ref": "#/components/schemas/CompletedCourseOut"
            },
            "title": "Completed Courses",
            "type": "array"
          },
          "lessons_finished": {
            "title": "Lessons Finished",
            "type": "integer"
          },
          "mastery": {
            "items": {
              "$ref": "#/components/schemas/MasteryOut"
            },
            "title": "Mastery",
            "type": "array"
          },
          "quizzes": {
            "items": {
              "$ref": "#/components/schemas/QuizHistoryOut"
            },
            "title": "Quizzes",
            "type": "array"
          }
        },
        "required": [
          "mastery",
          "completed_courses",
          "quizzes",
          "lessons_finished"
        ],
        "title": "RecordOut",
        "type": "object"
      },
...
      "StudentSignupIn": {
        "properties": {
          "board_code": {
            "maxLength": 10,
            "minLength": 1,
            "title": "Board Code",
            "type": "string"
          },
          "city": {
            "maxLength": 80,
            "minLength": 1,
            "title": "City",
            "type": "string"
          },
          "class_number": {
            "title": "Class Number",
            "type": "integer"
          },
          "email": {
            "anyOf": [
              {
                "maxLength": 254,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "maxLength": 150,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "mobile": {
            "anyOf": [
              {
                "maxLength": 254,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mobile"
          },
          "parent_contact": {
            "maxLength": 254,
            "minLength": 1,
            "title": "Parent Contact",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "school_name": {
            "default": "",
            "maxLength": 150,
            "title": "School Name",
            "type": "string"
          }
        },
        "required": [
          "full_name",
          "password",
          "class_number",
          "board_code",
          "city",
          "parent_contact"
        ],
        "title": "StudentSignupIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionAuth": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Django platform API used by the TUTOR web app.",
    "title": "TUTOR API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "apps_accounts_api_do_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Do Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/otp/send": {
      "post": {
        "operationId": "apps_accounts_api_otp_send",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpSendIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Otp Send",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "apps_accounts_api_otp_verify",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Otp Verify",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "operationId": "apps_accounts_api_password_reset",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Password Reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/signup/parent": {
      "post": {
        "operationId": "apps_accounts_api_signup_parent",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParentSignupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Signup Parent",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/signup/student": {
      "post": {
        "operationId": "apps_accounts_api_signup_student",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentSignupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Created"
          }
        },
        "summary": "Signup Student",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/consent/link/{token}": {
      "get": {
        "operationId": "apps_accounts_api_link_details",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentLinkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Link Details",
        "tags": [
          "consent"
        ]
      }
    },
    "/api/v1/consent/link/{token}/approve": {
      "post": {
        "operationId": "apps_accounts_api_approve",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "consent"
        ]
      }
    },
    "/api/v1/consent/link/{token}/report": {
      "post": {
        "operationId": "apps_accounts_api_report",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Report",
        "tags": [
          "consent"
        ]
      }
    },
    "/api/v1/consent/requests/parent-contact": {
      "patch": {
        "operationId": "apps_accounts_api_change_parent_contact",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParentContactIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Change Parent Contact",
        "tags": [
          "consent"
        ]
      }
    },
    "/api/v1/consent/requests/resend": {
      "post": {
        "operationId": "apps_accounts_api_resend",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Resend",
        "tags": [
          "consent"
        ]
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "apps_accounts_api_me",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/parent/children": {
      "post": {
        "operationId": "apps_accounts_api_add_child",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddChildIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildOut"
                }
              }
            },
            "description": "Created"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Add Child",
        "tags": [
          "parent"
        ]
      }
    },
    "/api/v1/parent/children/{child_id}/consent/restore": {
      "post": {
        "operationId": "apps_accounts_api_restore",
        "parameters": [
          {
            "in": "path",
            "name": "child_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Child Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Restore",
        "tags": [
          "parent"
        ]
      }
    },
    "/api/v1/parent/children/{child_id}/consent/withdraw": {
      "post": {
        "operationId": "apps_accounts_api_withdraw",
        "parameters": [
          {
            "in": "path",
            "name": "child_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Child Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Withdraw",
        "tags": [
          "parent"
        ]
      }
    },
...
    "/api/v1/student/record": {
      "get": {
        "description": "The student's own learner record. (Parents get their own view in the parent API.)",
        "operationId": "apps_learning_api_record",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Record",
        "tags": [
          "learning"
        ]
      }
    },
    "/api/v1/student/today": {
      "get": {
        "operationId": "apps_learning_api_today",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayOut"
                }
              }
            },
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "summary": "Today",
        "tags": [
          "learning"
        ]
      }
    }
  },
  "servers": []
}